+++ src/three/math/Ellipsoid.d.ts
@@ -1,10 +1,9 @@
 import { Vector3, Matrix4 } from 'three';
 
-export const ENU_FRAME: Number;
-export const CAMERA_FRAME: Number;
-export const OBJECT_FRAME: Number;
-
-type FRAME_TYPE = typeof ENU_FRAME | typeof CAMERA_FRAME | typeof OBJECT_FRAME;
+export enum Frames {}
+export const ENU_FRAME: Frames;
+export const CAMERA_FRAME: Frames;
+export const OBJECT_FRAME: Frames;
 
 export class Ellipsoid {
 
@@ -19,11 +18,11 @@ export class Ellipsoid {
 
 	getAzElRollFromRotationMatrix(
 		lat: Number, lon: Number, rotationMatrix: Matrix4,
-		target: Object, frame: FRAME_TYPE,
+		target: Object, frame: Frames,
 	): { azimuth: Number, elevation: Number, roll: Number };
 	getRotationMatrixFromAzElRoll(
 		lat: Number, lon: Number, az: Number, el: Number, roll: Number,
-		target: Matrix4, frame: FRAME_TYPE,
+		target: Matrix4, frame: Frames,
 	): Matrix4;
 
 	getEastNorthUpFrame( lat: Number, lon: Number, target: Matrix4 ): Matrix4;